+++ packages/react-google-maps-api/src/components/drawing/Polygon.tsx
@@ -210,7 +210,7 @@ function PolygonFunctional({
         google.maps.event.addListener(instance, 'dragend', onDragEnd)
       )
     }
-  }, [onDblClick])
+  }, [onDragEnd])
 
   useEffect(() => {
     if (instance && onDragStart) {